[JENKINS-76420] Disable the choice of deprecated plugin webhook management if bitbucket-webhook plugin is installed and active
If the bitbucket-webhook plugin is installed, the deprecated "Post Webhook for Jenkins" still displayed as available choice as webhook management in the Jenkins configuration page (Post Webhook for Jenkins, Native Data Center, Bitbucket API REST 1.0, Bitbucket API REST 2.0 (Experimental)).
