+++ src/main/resources/com/cloudbees/jenkins/plugins/bitbucket/impl/endpoint/BitbucketServerEndpoint/help-webhook.html
@@ -1,0 +1,10 @@
+<div>
+    <p>
+        By selecting a webhook management system, you can enable additional features specific to your
+        chosen implementation, such as automatic registration, incoming payload signature verification, and so on.
+        If you have access to the repositories, you can manually register your webhook using <a src="https://github.com/jenkinsci/bitbucket-branch-source-plugin/blob/master/docs/USER_GUIDE.adoc#manual-registration">this guide</a>
+    </p>
+    <p>
+        All incoming webhooks will be processed regardless of the chosen management;
+        the appropriate webhook processor will be selected based on the incoming payload.</p>
+</div>
